@@ -1,7 +1,7 @@
 require 'slimmer/headers'
 
 class Contact::GovukController < ContactController
-  before_filter proc {
+  before_action proc {
     hide_report_a_problem_form_in_response
   }, only: [:anonymous_feedback_thankyou, :named_contact_thankyou]
 
